requirements.txt
# Core dependencies
crewai==0.140.0
crewai-tools==0.49.0
langchain==0.3.26
langchain-cohere==0.3.5
langchain-community==0.3.26
langchain-core==0.3.67
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
openai==1.93.0

# Jira integration
jira==3.5.2
atlassian-python-api==3.41.4

# Vector database and embeddings
chromadb==0.4.24
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# Data processing
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.0

# CLI and UI
click==8.1.8
rich==13.7.1
typer==0.12.3

# Development tools
pytest==8.1.1
black==24.4.0
isort==5.13.2
mypy==1.10.0

# Optional: For future model training
transformers==4.40.1
torch==2.3.0
datasets==2.19.1
pydantic~=2.11.7
setuptools~=80.3.1